♻️ refactor :커뮤니티상세페이지 유틸 함수 폴더 이동
