+++ src/pages/CommunityDetail.tsx
@@ -17,7 +17,7 @@ import { LuArrowUpDown } from 'react-icons/lu'
 import CommentLoading from '../components/commnunityDetail/CommentLoading'
 import CommentTextArea from '../components/commnunityDetail/CommentTextArea'
 import { useSortComments } from '@hooks/useSortComments'
-import { URLCopy } from '@lib/index'
+import { URLCopy } from '@utils/formatDate'
 import { IoChatbubbleOutline } from 'react-icons/io5'
 import { fetchCommunityDetail } from '../api/community'
 import type { PostData } from '@customType/communityDetail'
